+++ lib/swid/frontend/dart/narrowModifierProducer.dart
